requirements/requirements.txt
apache-airflow==2.0.2
autoflake==1.4
black==20.8b1
google-api-core==1.26.2
google-api-python-client==1.12.8
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-cloud-bigquery==2.13.1
google-cloud-bigquery-storage==2.3.0
google-cloud-bigtable==1.7.0
google-cloud-container==2.3.1
google-cloud-core==1.6.0
google-cloud-dlp==3.0.1
google-cloud-language==2.0.0
google-cloud-secret-manager==2.3.0
google-cloud-spanner==3.2.0
google-cloud-speech==2.2.0
google-cloud-storage==1.36.2
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.2
google-cloud-videointelligence==2.0.0
google-cloud-vision==2.2.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
isort==5.8.0
pandas==1.0.0
pandas-gbq==0.13.2
pyarrow==3.0.0